pyproject.toml

[project]
name = "space_traders_wrapper"
version = "0.1.0"
description = "A wrapper for SpaceTraders API"
authors = [
    {name = "Daniel J Brown", email = "browndj3@gmail.com"},
]
dependencies = [
    "pydantic>=2.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0.post2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.291",
    "black>=23.9.1",
]
